Feat: Change Byte

+++ src/Application/Common/Dtos/AppointmentDto.cs
@@ -4,7 +4,7 @@ public sealed record AppointmentDto : BaseDto
 {
     public string? Description { get; set; }
     public DateTime KeepDate { get; set; }
-    public byte AmountHours { get; set; }
+    public int AmountHours { get; set; }
     public Guid AssignmentId { get; set; }
     public Guid UserId { get; set; }
