+++ src/Domain/Entities/Assignment.cs
@@ -6,7 +6,7 @@ public sealed class Assignment : BaseEntity
     public string? Description { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
-    public byte AmountHours { get; set; }
+    public int AmountHours { get; set; }
 
     public Guid ProjectId { get; set; }
     public Project? Project { get; set; }
